pyproject.toml
[build-system]
requires = ["poetry>=1.0.2"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "imctools"
version = "2.0.0"
description = "Tools to handle IMC data"
license = "LICENSE"
authors = [
    "Vito Zanotelli <vito.zanotelli@uzh.ch>",
    "Anton Rau <anton.rau@uzh.ch>"
]
homepage = "https://github.com/BodenmillerGroup/imctools"
repository = "https://github.com/BodenmillerGroup/imctools"
documentation = "https://imctools.readthedocs.io"
keywords = ["mcd", "ome-tiff", "imc", "cellprofiler", "fiji"]
readme = "README.md"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.7",
]
include = ["AUTHORS.md", "CONTRIBUTING.md", "CHANGELOG.md", "LICENSE", "README.md"]
exclude = ["tests", "docs"]

[tool.poetry.dependencies]
python = ">=3.7"
pandas = ">=1.0.0rc0"
pyyaml = "^5.3"
xmltodict = ">=0.12.0"
xtiff = ">=0.5.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.3.1"
black = "^19.10b0"
bump2version = "^1.0.0"
flake8 = "^3.7.9"
isort = "^4.3.21"
mypy = "*"
pdoc3 = "^0.7.4"
pylint = "^2.4.4"
pytest = "^5.3.4"
pytest-cov = "^2.8.1"
vulture = "^1.2"

[tool.black]
line-length = 120
target-version = [ "py37" ]
